requirements.txt
Django>=2.2
python-dateutil
requests
